Call set_curr_task() when needed
In some cases, the scheduler invokes set_curr_task() before enqueueing the
task. But if the task is enqueued as RT, it can be demoted during enqueue...
In this case, set_curr_task() is called for the RT scheduling class, but
the task ends up being enqueued in the CFS rq... And set_curr_task() is
not invoked for the CFS scheduling class!

Fix this by explicitly invoking the CFS set_curr_task() (if needed) in
case of demotion, before enqueueing in the CFS rq.