Mysterious changes to improve the patchset stability
The demotion mechanism currently still has some bugs, that can be
triggered by using the "periodic1.sh" or "periodic2.sh" scripts.
After some experiments, it turned out that these changes improve the
stability of the patchset (with this patch, the demotion mechanism can
survive to 33mins of "periodic1.sh" or "periodic2.sh".
The bugs are probably still there, though.